+++ relay-monitors/src/lib.rs
@@ -114,6 +114,13 @@ pub struct MonitorConfig {
     /// How many consecutive OK check-ins it takes to resolve an issue.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     recovery_threshold: Option<u64>,
+
+    /// Who the owner of the monitor should be. Uses the ActorTuple [0]
+    /// identifier format.
+    ///
+    /// [0]: https://github.com/getsentry/sentry/blob/3644f5c4f2a99073bf925181b5237a6e05c1d6c2/src/sentry/utils/actor.py#L17
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    owner: Option<String>,
 }
 
 /// The trace context sent with a check-in.
@@ -325,7 +332,8 @@ mod tests {
     "max_runtime": 10,
     "timezone": "America/Los_Angles",
     "failure_issue_threshold": 3,
-    "recovery_threshold": 1
+    "recovery_threshold": 1,
+    "owner": "user:123"
   }
 }"#;