+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 **Features**:
 
 - Use same keys for OTel span attributes and Sentry span data. ([#3457](https://github.com/getsentry/relay/pull/3457))
+- Support passing owner when upserting Monitors. ([#3468](https://github.com/getsentry/relay/pull/3468))
 
 **Internal**: